Added PDFs

+++ Uebung_9-Procedures_vs_Call_Backs/04_traffic_light.cpp
@@ -19,7 +19,7 @@
 class TL {
     private:
         const char *light;
-        int phase;
+        short int phase;
 
         void change_light ();
